create debug submodule
