+++ packages/react-native-client/src/debug/index.ts
@@ -1,0 +1,4 @@
+export { useRTCStatistics } from './stats/useRTCStatistics';
+
+export type { LoggingSeverity } from './common/webRTC';
+export { changeWebRTCLoggingSeverity } from './common/webRTC';
